scripts/requirements.txt
google-cloud-storage >= 2.7.0
urllib3 >= 1.26.13